Alomst there,woo-hoo!

+++ SB_Down/Chassis/Core/Inc/can.h
@@ -42,7 +42,7 @@ void can_cmd_send(int motor1,int motor2,int motor3,int motor4);
 void can_remote(uint8_t sbus_buf[],uint8_t can_send_id);
 typedef struct
 {
-	uint16_t angle;
+	double angle;
 	int16_t speed;
 	uint16_t tor_current;
 	uint16_t temperture;
