+++ SB_Down/Chassis/USER/handle_value.h
@@ -2,7 +2,7 @@
 #define _HANDLE_VALUE_H_
 #include "main.h"
 
-int16_t remote_value(int16_t x, int16_t y);
+float remote_value(int16_t x, int16_t y);
 int16_t motor_value(int16_t get, int16_t set);
 
 #endif
